Upgrade for Ubuntu14.04HVM and Chef-Server 11.4.1

+++ files/knife.rb
@@ -1,11 +1,12 @@
-system('sudo install -o ubuntu -g ubuntu -m 0600 /etc/chef-server/admin.pem ~/.chef/')
-system('sudo install -o ubuntu -g ubuntu -m 0600 /etc/chef-server/chef-validator.pem ~/.chef/')
+system('sudo install -o ubuntu -g ubuntu -m 0600 /etc/chef-server/admin.pem ~/.chef/') unless File.exists?(File.expand_path('../admin.pem', __FILE__))
+system('sudo install -o ubuntu -g ubuntu -m 0600 /etc/chef-server/chef-validator.pem ~/.chef/') unless File.exists?(File.expand_path('../chef-validator.pem', __FILE__))
+
 
 require 'ohai'
 ohai=Ohai::System.new
 ohai.all_plugins
 
-server_name = ohai.cloud[:public_ipv4]
+server_name = ohai[:cloud][:public_ipv4]
 
 log_level                :info
 log_location             STDOUT
